[Pal/Linux] Move x86_64-specific signal funcs into arch-specific file
